Refactor the code for the vector field calculation
