+++ src/DataStream/VectorField.cc
@@ -1,0 +1,62 @@
+/* This file is part of the CARTA Image Viewer: https://github.com/CARTAvis/carta-backend
+   Copyright 2018-2022 Academia Sinica Institute of Astronomy and Astrophysics (ASIAA),
+   Associated Universities, Inc. (AUI) and the Inter-University Institute for Data Intensive Astronomy (IDIA)
+   SPDX-License-Identifier: GPL-3.0-or-later
+*/
+
+#include "VectorField.h"
+
+namespace carta {
+
+void GetTiles(int image_width, int image_height, int mip, std::vector<Tile>& tiles) {
+    int tile_size_original = TILE_SIZE * mip;
+    int num_tile_columns = ceil((double)image_width / tile_size_original);
+    int num_tile_rows = ceil((double)image_height / tile_size_original);
+    int32_t tile_layer = Tile::MipToLayer(mip, image_width, image_height, TILE_SIZE, TILE_SIZE);
+    tiles.resize(num_tile_rows * num_tile_columns);
+
+    for (int j = 0; j < num_tile_rows; ++j) {
+        for (int i = 0; i < num_tile_columns; ++i) {
+            tiles[j * num_tile_columns + i].x = i;
+            tiles[j * num_tile_columns + i].y = j;
+            tiles[j * num_tile_columns + i].layer = tile_layer;
+        }
+    }
+}
+
+void FillTileData(CARTA::TileData* tile, int32_t x, int32_t y, int32_t layer, int32_t mip, int32_t tile_width, int32_t tile_height,
+    std::vector<float>& array, CARTA::CompressionType compression_type, float compression_quality) {
+    if (tile) {
+        tile->set_x(x);
+        tile->set_y(y);
+        tile->set_layer(layer);
+        tile->set_mip(mip);
+        tile->set_width(tile_width);
+        tile->set_height(tile_height);
+        if (compression_type == CARTA::CompressionType::ZFP) {
+            // Get and fill the NaN data
+            auto nan_encodings = GetNanEncodingsBlock(array, 0, tile_width, tile_height);
+            tile->set_nan_encodings(nan_encodings.data(), sizeof(int32_t) * nan_encodings.size());
+            // Compress and fill the data
+            std::vector<char> compression_buffer;
+            size_t compressed_size;
+            int precision = lround(compression_quality);
+            Compress(array, 0, compression_buffer, compressed_size, tile_width, tile_height, precision);
+            tile->set_image_data(compression_buffer.data(), compressed_size);
+        } else {
+            tile->set_image_data(array.data(), sizeof(float) * array.size());
+        }
+    }
+}
+
+CARTA::ImageBounds GetImageBounds(const Tile& tile, int image_width, int image_height, int mip) {
+    int tile_size_original = TILE_SIZE * mip;
+    CARTA::ImageBounds bounds;
+    bounds.set_x_min(std::min(std::max(0, tile.x * tile_size_original), image_width));
+    bounds.set_x_max(std::min(image_width, (tile.x + 1) * tile_size_original));
+    bounds.set_y_min(std::min(std::max(0, tile.y * tile_size_original), image_height));
+    bounds.set_y_max(std::min(image_height, (tile.y + 1) * tile_size_original));
+    return bounds;
+}
+
+} // namespace carta
